// src/lib/constant.d.ts
export declare enum Color {
    PaleRed = "#FF7256",
    ShallowRed = "#FF6347",
    LightRed = "#FF4500",
    Red = "#FF0000",
    HeavyRed = "#EE0000",
    DeepRed = "#CD0000",
    DarkRed = "#8B0000",
    PaleBlue = "#81D4FA",
    ShallowBlue = "#00BFFF",
    LightBlue = "#1E90FF",
    Blue = "#0000FF",
    HeavyBlue = "#0000CD",
    DeepBlue = "#0000AC",
    DarkBlue = "#00008B",
    PaleYellow = "#FFFFE0",
    ShallowYellow = "#FFF68F",
    LightYellow = "#FFF68F",
    Yellow = "#FFFF00",
    HeavyYellow = "#EEEE00",
    DeepYellow = "#EEC900",
    DarkYellow = "#CDCD00",
    PaleGreen = "#9AFF9A",
    ShallowGreen = "#00FF7F",
    LightGreen = "#7FFF00",
    Green = "#00FF00",
    HeavyGreen = "#66CD00",
    DeepGreen = "#2E8B57",
    DarkGreen = "#006400",
    PaleGray = "#E8E8E8",
    ShallowGray = "#CFCFCF",
    LightGray = "#B5B5B5",
    Gray = "#9C9C9C",
    HeavyGray = "#828282",
    DeepGray = "#4F4F4F",
    DarkGray = "#363636",
    Dark = "#1C1C1C",
    Black = "#000000",
    Orange = "#FFA500",
    Azure = "#FFA500",
    Gold = "#FFD700",
    Pink = "#FFC0CB",
    White = "#FFFFFF",
    /** 米黄色 */
    Beige = "#F7EED6",
    /** 花卉白 */
    FloralWhite = "#FFFAF0"
}
export declare enum Shadow {
    Base = "0px 2px 5px 0px rgba(0, 0, 0, 0.1)",
    Pale = "0px 2px 5px 1px #0000001e"
}